+++ FitzyBot.Core/Entities/Person.cs
@@ -5,6 +5,6 @@ namespace FitzyBot.Core
     public class Person
     {        
         public string TwitchUsername { get; set; }
-        public decimal Balance { get; set; }
+        public int Balance { get; set; }
     }
 }
